requirements.txt
flask
openai
gitpython
langchain
faiss-cpu
requests